TZ Fix and Message for deleting accounts

+++ src/pages/accounts/accounts.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { ModalController, ActionSheetController } from 'ionic-angular';
+import { ModalController, AlertController, ActionSheetController } from 'ionic-angular';
 import { AccountModalComponent } from '../../components/account-modal/account-modal';
 import { AccountsProvider } from '../../providers/accounts/accounts';
 import { Events } from 'ionic-angular';
@@ -16,7 +16,7 @@ export class AccountsPage {
 
   accounts: any = [];
 
-  constructor(public modalCtrl: ModalController, public alertCtrl: ActionSheetController, public accountsProvider: AccountsProvider,
+  constructor(public modalCtrl: ModalController, public alertDialogCtrl: AlertController, public alertCtrl: ActionSheetController, public accountsProvider: AccountsProvider,
     public events: Events, public sorter: SorterProvider, public transactionsProvider: TransactionsProvider, public categoriesProvider: CategoriesProvider,
     public moneyProvider: MoneyProvider) {
     this.refreshAccounts();
@@ -55,43 +55,64 @@ export class AccountsPage {
           text: 'Delete',
           role: 'destructive',
           handler: () => {
-            this.accountsProvider.deleteAccount(account._id).then(() => {
-              return this.moneyProvider.getAvailableToBudget().then(
-                (atob) => {
-                  let newTransactionData = {
-                    name: "Closing account: " + account.name,
-                    status: "Cleared",
-                    date: (new Date()).toISOString(),
-                    amount: Number(account.amount) / 100,
-                    direction: null,
-                    visible: false,
-                    sourceCatId: null, //{ sourceCatId: null, sourceCatName: null },
-                    sourceAcctId: null, //{ sourceAcctId: null, sourceAcctName: null },
-                    destCatId: null, //{ destCatId: null, destCatName: null },
-                    destAcctId: null, //{ destAcctId: null, destAcctName: null },
-                  }
 
-                  if (account.amount >= 0) {
-                    newTransactionData.direction = "Outflow";
-                    newTransactionData.sourceCatId = atob._id;
-                  } else if (account.amount < 0) {
-                    newTransactionData.amount *= -1;
-                    newTransactionData.direction = "Inflow";
-                    newTransactionData.destCatId = atob._id;
+            let alert = this.alertDialogCtrl.create({
+              title: 'Delete Account?',
+              message: "Your Available to Budget amount will be updated to reflect this account being removed.",
+              buttons: [
+                {
+                  text: 'Cancel',
+                  role: 'cancel',
+                  handler: () => {
+                    console.log('Cancel clicked');
                   }
+                },
+                {
+                  text: 'Delete',
+                  role: 'destructive',
+                  handler: () => {
+                    this.accountsProvider.deleteAccount(account._id).then(() => {
+                      return this.moneyProvider.getAvailableToBudget().then(
+                        (atob) => {
+                          let newTransactionData = {
+                            name: "Closing account: " + account.name,
+                            status: "Cleared",
+                            date: this.transactionsProvider.getNewDate(),
+                            amount: Number(account.amount) / 100,
+                            direction: null,
+                            visible: false,
+                            sourceCatId: null, //{ sourceCatId: null, sourceCatName: null },
+                            sourceAcctId: null, //{ sourceAcctId: null, sourceAcctName: null },
+                            destCatId: null, //{ destCatId: null, destCatName: null },
+                            destAcctId: null, //{ destAcctId: null, destAcctName: null },
+                          }
 
-                  return this.transactionsProvider.addTransaction(newTransactionData).then(
-                    () => {
-                      return this.moneyProvider.getCategoryTotal(atob._id).then(
-                        (catTotal) => {
-                          return this.moneyProvider.updateCategoryAmount(atob._id, catTotal).then(
+                          if (account.amount >= 0) {
+                            newTransactionData.direction = "Outflow";
+                            newTransactionData.sourceCatId = atob._id;
+                          } else if (account.amount < 0) {
+                            newTransactionData.amount *= -1;
+                            newTransactionData.direction = "Inflow";
+                            newTransactionData.destCatId = atob._id;
+                          }
+
+                          return this.transactionsProvider.addTransaction(newTransactionData).then(
                             () => {
-                              this.events.publish('account:deleted');
+                              return this.moneyProvider.getCategoryTotal(atob._id).then(
+                                (catTotal) => {
+                                  return this.moneyProvider.updateCategoryAmount(atob._id, catTotal).then(
+                                    () => {
+                                      this.events.publish('account:deleted');
+                                    });
+                                });
                             });
                         });
                     });
-                });
+                  }
+                }
+              ]
             });
+            alert.present();
           }
         },
         {
